Extract most of `cid-fmt` logic so it can be used as a library.
