+++ cid-fmt/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"os"
 	"strings"
@@ -9,37 +8,13 @@ import (
 	c "github.com/ipfs/go-cid"
 
 	mb "github.com/multiformats/go-multibase"
-	mh "github.com/multiformats/go-multihash"
 )
 
 func usage() {
 	fmt.Fprintf(os.Stderr, "usage: %s [-b multibase-code] [-v cid-version] <fmt-str> <cid> ...\n\n", os.Args[0])
-	fmt.Fprintf(os.Stderr, "<fmt-str> is either 'prefix' or a printf style format string:\n%s", fmtRef)
+	fmt.Fprintf(os.Stderr, "<fmt-str> is either 'prefix' or a printf style format string:\n%s", c.FormatRef)
 	os.Exit(2)
 }
-
-const fmtRef = `
-   %% literal %
-   %b multibase name 
-   %B multibase code
-   %v version string
-   %V version number
-   %c codec name
-   %C codec code
-   %h multihash name
-   %H multihash code
-   %L hash digest length
-   %m multihash encoded in base %b (with multibase prefix)
-   %M multihash encoded in base %b without multibase prefix
-   %d hash digest encoded in base %b (with multibase prefix)
-   %D hash digest encoded in base %b without multibase prefix
-   %s cid string encoded in base %b (1)
-   %S cid string encoded in base %b without multibase prefix
-   %P cid prefix: %v-%c-%h-%L
-
-(1) For CID version 0 the multibase must be base58btc and no prefix is
-used.  For Cid version 1 the multibase prefix is included.
-`
 
 func main() {
 	if len(os.Args) < 2 {
@@ -94,7 +69,7 @@ outer:
 		}
 	}
 	for _, cidStr := range args[1:] {
-		base, cid, err := decode(cidStr)
+		base, cid, err := c.DecodeV2(cidStr)
 		if err != nil {
 			fmt.Fprintf(os.Stdout, "!INVALID_CID!\n")
 			errorMsg("%s: %v", cidStr, err)
@@ -113,9 +88,9 @@ outer:
 				continue
 			}
 		}
-		str, err := fmtCid(fmtStr, base, cid)
+		str, err := c.Format(fmtStr, base, cid)
 		switch err.(type) {
-		case FormatStringError:
+		case c.FormatStringError:
 			fmt.Fprintf(os.Stderr, "Error: %v\n", err)
 			os.Exit(2)
 		default:
@@ -138,146 +113,6 @@ func errorMsg(fmtStr string, a ...interface{}) {
 	fmt.Fprintf(os.Stderr, fmtStr, a...)
 	fmt.Fprintf(os.Stderr, "\n")
 	exitCode = 1
-}
-
-func decode(v string) (mb.Encoding, *c.Cid, error) {
-	if len(v) < 2 {
-		return 0, nil, c.ErrCidTooShort
-	}
-
-	if len(v) == 46 && v[:2] == "Qm" {
-		hash, err := mh.FromB58String(v)
-		if err != nil {
-			return 0, nil, err
-		}
-
-		return mb.Base58BTC, c.NewCidV0(hash), nil
-	}
-
-	base, data, err := mb.Decode(v)
-	if err != nil {
-		return 0, nil, err
-	}
-
-	cid, err := c.Cast(data)
-
-	return base, cid, err
-}
-
-const ERR_STR = "!ERROR!"
-
-type FormatStringError struct {
-	Message   string
-	Specifier string
-}
-
-func (e FormatStringError) Error() string {
-	if e.Specifier == "" {
-		return e.Message
-	} else {
-		return fmt.Sprintf("%s: %s", e.Message, e.Specifier)
-	}
-}
-
-func fmtCid(fmtStr string, base mb.Encoding, cid *c.Cid) (string, error) {
-	p := cid.Prefix()
-	out := new(bytes.Buffer)
-	var err error
-	encoder, err := mb.NewEncoder(base)
-	if err != nil {
-		return "", err
-	}
-	for i := 0; i < len(fmtStr); i++ {
-		if fmtStr[i] != '%' {
-			out.WriteByte(fmtStr[i])
-			continue
-		}
-		i++
-		if i >= len(fmtStr) {
-			return "", FormatStringError{"premature end of format string", ""}
-		}
-		switch fmtStr[i] {
-		case '%':
-			out.WriteByte('%')
-		case 'b': // base name
-			out.WriteString(baseToString(base))
-		case 'B': // base code
-			out.WriteByte(byte(base))
-		case 'v': // version string
-			fmt.Fprintf(out, "cidv%d", p.Version)
-		case 'V': // version num
-			fmt.Fprintf(out, "%d", p.Version)
-		case 'c': // codec name
-			out.WriteString(codecToString(p.Codec))
-		case 'C': // codec code
-			fmt.Fprintf(out, "%d", p.Codec)
-		case 'h': // hash fun name
-			out.WriteString(hashToString(p.MhType))
-		case 'H': // hash fun code
-			fmt.Fprintf(out, "%d", p.MhType)
-		case 'L': // hash length
-			fmt.Fprintf(out, "%d", p.MhLength)
-		case 'm', 'M': // multihash encoded in base %b
-			out.WriteString(encode(encoder, cid.Hash(), fmtStr[i] == 'M'))
-		case 'd', 'D': // hash digest encoded in base %b
-			dec, err := mh.Decode(cid.Hash())
-			if err != nil {
-				return "", err
-			}
-			out.WriteString(encode(encoder, dec.Digest, fmtStr[i] == 'D'))
-		case 's': // cid string encoded in base %b
-			str, err := cid.StringOfBase(base)
-			if err != nil {
-				return "", err
-			}
-			out.WriteString(str)
-		case 'S': // cid string without base prefix
-			out.WriteString(encode(encoder, cid.Bytes(), true))
-		case 'P': // prefix
-			fmt.Fprintf(out, "cidv%d-%s-%s-%d",
-				p.Version,
-				codecToString(p.Codec),
-				hashToString(p.MhType),
-				p.MhLength,
-			)
-		default:
-			return "", FormatStringError{"unrecognized specifier in format string", fmtStr[i-1 : i+1]}
-		}
-
-	}
-	return out.String(), err
-}
-
-func baseToString(base mb.Encoding) string {
-	baseStr, ok := mb.EncodingToStr[base]
-	if !ok {
-		return fmt.Sprintf("base?%c", base)
-	}
-	return baseStr
-}
-
-func codecToString(num uint64) string {
-	name, ok := c.CodecToStr[num]
-	if !ok {
-		return fmt.Sprintf("codec?%d", num)
-	}
-	return name
-}
-
-func hashToString(num uint64) string {
-	name, ok := mh.Codes[num]
-	if !ok {
-		return fmt.Sprintf("hash?%d", num)
-	}
-	return name
-}
-
-func encode(base mb.Encoder, data []byte, strip bool) string {
-	str := base.Encode(data)
-	if strip {
-		return str[1:]
-	}
-	return str
 }
 
 func toCidV0(cid *c.Cid) (*c.Cid, error) {
